update the graph block to handle the new metrics
