--- plugin/blocks/src/components/graph-v2/block.json
@@ -1,41 +1,0 @@
-{
-	"$schema": "https://schemas.wp.org/trunk/block.json",
-	"apiVersion": 3,
-	"name": "wpcloud/graph-v2",
-	"version": "0.1.0",
-	"title": "graph v2",
-	"category": "wpcloud",
-	"icon": "chart-line",
-	"description": "Graph block for WPCloud plot blocks",
-	"example": {},
-	"supports": {
-		"html": false
-	},
-	"textdomain": "graph",
-	"editorScript": "file:./index.js",
-	"editorStyle": "file:./index.css",
-	"style": "file:./style-index.css",
-	"ancestor": ["wpcloud/metrics"],
-	"attributes" : {
-		"metric" : {
-			"type" : "string",
-			"default" : "requests"
-		},
-		"dimension" : {
-			"type" : "string",
-			"default" : "https_status"
-		},
-		"type": {
-			"type": "string",
-			"default": "line"
-		},
-		"title" : {
-			"type" : "string",
-			"default" : "Status Codes"
-		},
-		"showLegend" : {
-			"type" : "boolean",
-			"default" : true
-		}
-	}
-}
