--- plugin/blocks/src/components/graph-v2/components/loader.js
@@ -1,20 +1,0 @@
-
-/**
- * WordPress dependencies
- */
-import { Spinner } from '@wordpress/components';
-
-export default () => {
-	return (
-		<>
-			<div style={{ width: "100%", height: "100%", backgroundColor: "white", opacity: "66%" }} />
-			<div style={{ position: "absolute", top: "50%", left: "50%", transform: "translate(-50%, -50%)" }}>
-				<Spinner  style={{
-  					height: 'calc(4px * 20)',
-    				width: 'calc(4px * 20)'
-				}}
-				/>
-			</div>
-		</>
-	);
-}
