--- plugin/blocks/src/components/graph-v2/index.js
@@ -1,26 +1,0 @@
-
-/**
- * WordPress dependencies
- */
-import { registerBlockType } from '@wordpress/blocks';
-import { useBlockProps } from '@wordpress/block-editor';
-import './style.scss';
-
-/**
- * Internal dependencies
- */
-import Edit from './edit';
-import metadata from './block.json';
-
-
-registerBlockType( metadata.name, {
-	/**
-	 * @see ./edit.js
-	 */
-	edit: Edit,
-
-	/**
-	 * @see ./save.js
-	 */
-	save: ({ attributes }) => (<div {...useBlockProps.save()} data-graph-attributes={JSON.stringify(attributes)} >Graph</div>)
-} );
