--- plugin/blocks/src/components/graph-v2/index.php
@@ -1,3 +1,0 @@
-<?php
-
-// @TODO: don't forget that the uplot styles are being enqueued in the old graph block
