+++ plugin/blocks/src/components/graph/edit.js
@@ -1,4 +1,5 @@
 
+import { useEffect, useState } from 'react';
 /**
  * WordPress dependencies
  */
@@ -13,47 +14,74 @@ import {
 	TextControl,
 	ToggleControl,
 	SelectControl,
+	Spinner
 } from '@wordpress/components';
-import { useCallback, useEffect } from '@wordpress/element';
 
+import apiFetch from '@wordpress/api-fetch';
+
+import {  updateAttribute } from '@wpcloud/controls';
 import './editor.scss';
 
 
 export default function Edit( { attributes, setAttributes } ) {
 
-	const { metric, type, title, showLegend } = attributes;
+	const { metric, dimension, type, title, showLegend } = attributes;
+	const update = updateAttribute(setAttributes);
+
+	const [metrics, setMetrics] = useState({});
+	const [dimensions, setDimensions] = useState({});
+	const [loading, setLoading] = useState(true);
+
+	useEffect(() => {
+		async function fetchAvailable() {
+			const available = await apiFetch({ path: 'wpcloud-station/v1/metrics/available' });
+			setMetrics(available.metrics);
+			setDimensions(available.dimensions);
+			setLoading(false);
+		}
+		fetchAvailable();
+	}, []);
 
 	const controls = (
 		<InspectorControls>
 			<PanelBody title={__('graph Settings')}>
 
-				<SelectControl
-					label={ __( 'Metric' ) }
-					value={ metric }
-					options={ [
-						{ label: __( 'Status Codes' ), value: 'status_codes' },
-					] }
-					onChange={(newVal) => setAttributes({ metric: newVal })}
-				/>
+				{loading && <Spinner />}
+				{!loading && (
+					<>
+						<SelectControl
+							label={__('Metric')}
+							value={metric}
+							options={Object.keys(metrics).map((key) => ({ label: metrics[key], value: key })) }
+							onChange={update('metric')}
+						/>
+						<SelectControl
+							label={__('Dimension')}
+							value={dimension}
+							options={Object.keys(dimensions).map((key) => ({ label: dimensions[key], value: key })) }
+							onChange={update('dimension')}
+						/>
+					</>)}
 				<SelectControl
 					label={ __( 'Type' ) }
 					value={ type }
 					options={ [
 						{ label: __( 'Area' ), value: 'area' },
 						{ label: __( 'Line' ), value: 'line' },
-						{ label: __( 'Bar' ), value: 'bar' },
+						{ label: __('Bar'), value: 'bar' },
+						{ label: __('Stacked Bar'), value: 'stacked-bar' },
 					] }
-					onChange={ ( newVal ) => setAttributes( { type: newVal } ) }
+					onChange={ update('type') }
 				/>
 				<TextControl
 					label={ __( 'Title' ) }
 					value={ title }
-					onChange={ ( newVal ) => setAttributes( { title: newVal } ) }
+					onChange={ update('title') }
 				/>
 				<ToggleControl
 					label={ __( 'Show Legend' ) }
 					checked={ showLegend }
-					onChange={ ( newVal ) => setAttributes( { showLegend: newVal } ) }
+					onChange={ update('showLegend') }
 				/>
 			</PanelBody>
 		</InspectorControls>
